face-rec/requirements.txt
Flask==2.2.5
Flask-Cors==5.0.0
requests==2.31.0
opencv-python-headless==4.5.5.64
tensorflow==2.2.0
keras==2.3.1
mtcnn==0.1.0
scikit-learn==0.24.2
numpy==1.19.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==20.1.0
h5py==2.10.0